refactoring: materials
