+++ src/main/kotlin/com/github/trcdevelopers/clayium/common/recipe/builder/RecipeBuilder.kt
@@ -10,7 +10,7 @@ import com.github.trcdevelopers.clayium.common.recipe.ingredient.CRecipeInput
 import com.github.trcdevelopers.clayium.common.recipe.registry.RecipeRegistry
 import com.github.trcdevelopers.clayium.common.unification.OreDictUnifier
 import com.github.trcdevelopers.clayium.common.unification.OrePrefix
-import com.github.trcdevelopers.clayium.common.unification.material.Material
+import com.github.trcdevelopers.clayium.common.unification.material.EnumMaterial
 import com.github.trcdevelopers.clayium.common.unification.stack.UnificationEntry
 import net.minecraft.block.Block
 import net.minecraft.item.Item
@@ -53,7 +53,7 @@ abstract class RecipeBuilder<R: RecipeBuilder<R>>(
     fun input(metaItem: MetaItemClayium.MetaValueItem, amount: Int = 1) = input(metaItem.getStackForm(amount))
     fun input(block: Block, amount: Int = 1) = input(ItemStack(block, amount))
     fun input(oreDict: String, amount: Int = 1) = inputs(COreRecipeInput(oreDict, amount))
-    fun input(orePrefix: OrePrefix, material: Material, amount: Int = 1) = inputs(COreRecipeInput(UnificationEntry(orePrefix, material).toString(), amount))
+    fun input(orePrefix: OrePrefix, material: EnumMaterial, amount: Int = 1) = inputs(COreRecipeInput(UnificationEntry(orePrefix, material).toString(), amount))
 
     fun outputs(vararg stacks: ItemStack): R {
         outputs.addAll(stacks)
@@ -65,7 +65,7 @@ abstract class RecipeBuilder<R: RecipeBuilder<R>>(
     fun output(metaItem: MetaItemClayium.MetaValueItem, amount: Int = 1) = output(metaItem.getStackForm(amount))
     fun output(block: Block, amount: Int = 1) = output(ItemStack(block, amount))
     fun output(oreDict: String, amount: Int = 1) = outputs(OreDictUnifier.get(oreDict, amount))
-    fun output(orePrefix: OrePrefix, material: Material, amount: Int = 1) = outputs(OreDictUnifier.get(orePrefix, material, amount))
+    fun output(orePrefix: OrePrefix, material: EnumMaterial, amount: Int = 1) = outputs(OreDictUnifier.get(orePrefix, material, amount))
 
     fun duration(duration: Int) = duration(duration.toLong())
     fun duration(duration: Long): R {
