+++ src/main/kotlin/com/github/trcdevelopers/clayium/common/recipe/handler/MaterialRecipeHandler.kt
@@ -1,0 +1,14 @@
+package com.github.trcdevelopers.clayium.common.recipe.handler
+
+import com.github.trcdevelopers.clayium.common.unification.OrePrefix
+import com.github.trcdevelopers.clayium.common.unification.material.EnumMaterial
+
+object MaterialRecipeHandler {
+    fun registerRecipes() {
+
+    }
+
+    private fun handleIngot(orePrefix: OrePrefix, material: EnumMaterial) {
+
+    }
+}
