+++ src/main/kotlin/com/github/trcdevelopers/clayium/common/recipe/loader/BendingMachineLoader.kt
@@ -3,7 +3,7 @@ package com.github.trcdevelopers.clayium.common.recipe.loader
 import com.github.trcdevelopers.clayium.common.recipe.registry.CRecipes
 import com.github.trcdevelopers.clayium.common.unification.OreDictUnifier
 import com.github.trcdevelopers.clayium.common.unification.OrePrefix
-import com.github.trcdevelopers.clayium.common.unification.material.Material
+import com.github.trcdevelopers.clayium.common.unification.material.EnumMaterial
 import com.github.trcdevelopers.clayium.common.unification.material.MaterialProperty.Ingot
 import com.github.trcdevelopers.clayium.common.unification.material.MaterialProperty.Matter
 import com.github.trcdevelopers.clayium.common.unification.material.MaterialProperty.Plate
@@ -12,7 +12,7 @@ object BendingMachineLoader {
     fun register() {
         val registry = CRecipes.BENDING
 
-        for (material in Material.entries) {
+        for (material in EnumMaterial.entries) {
             if (material.hasProperty<Ingot>() && material.hasProperty<Plate>()) {
                 val plateProp = material.getProperty<Plate>() ?: continue
                 registry.register {
@@ -41,7 +41,7 @@ object BendingMachineLoader {
         }
     }
 
-    private fun registerLargePlateRecipe(material: Material) {
+    private fun registerLargePlateRecipe(material: EnumMaterial) {
         //todo
         // recipe duration is not always normalPlate * 4
     }
