+++ src/main/kotlin/com/github/trcdevelopers/clayium/common/recipe/loader/CaCondenserRecipeLoader.kt
@@ -4,13 +4,13 @@ import com.github.trcdevelopers.clayium.common.clayenergy.ClayEnergy
 import com.github.trcdevelopers.clayium.common.items.metaitem.MetaItemClayParts
 import com.github.trcdevelopers.clayium.common.recipe.registry.CRecipes
 import com.github.trcdevelopers.clayium.common.unification.OrePrefix
-import com.github.trcdevelopers.clayium.common.unification.material.Material
+import com.github.trcdevelopers.clayium.common.unification.material.EnumMaterial
 
 object CaCondenserRecipeLoader {
     fun registerRecipes() {
         CRecipes.CA_CONDENSER.register {
             input(MetaItemClayParts.ANTIMATTER_SEED)
-            output(OrePrefix.MATTER, Material.ANTIMATTER)
+            output(OrePrefix.MATTER, EnumMaterial.ANTIMATTER)
             cePerTick(ClayEnergy.of(250))
             duration(2000)
             tier(0)
