+++ src/main/kotlin/com/github/trcdevelopers/clayium/common/recipe/loader/ClayBlastFurnaceRecipeLoader.kt
@@ -3,7 +3,7 @@ package com.github.trcdevelopers.clayium.common.recipe.loader
 import com.github.trcdevelopers.clayium.common.clayenergy.ClayEnergy
 import com.github.trcdevelopers.clayium.common.recipe.registry.CRecipes
 import com.github.trcdevelopers.clayium.common.unification.OrePrefix
-import com.github.trcdevelopers.clayium.common.unification.material.Material
+import com.github.trcdevelopers.clayium.common.unification.material.EnumMaterial
 import net.minecraft.init.Items
 
 object ClayBlastFurnaceRecipeLoader {
@@ -13,7 +13,7 @@ object ClayBlastFurnaceRecipeLoader {
         registry.register {
             input(Items.IRON_INGOT)
             input(Items.COAL, 2)
-            output(OrePrefix.INGOT, Material.STEEL)
+            output(OrePrefix.INGOT, EnumMaterial.STEEL)
             cePerTick(ClayEnergy.milli(100))
             duration(500)
             tier(6)
