+++ src/main/kotlin/com/github/trcdevelopers/clayium/common/recipe/loader/CondenserRecipeLoader.kt
@@ -3,7 +3,7 @@ package com.github.trcdevelopers.clayium.common.recipe.loader
 import com.github.trcdevelopers.clayium.common.clayenergy.ClayEnergy
 import com.github.trcdevelopers.clayium.common.recipe.registry.CRecipes
 import com.github.trcdevelopers.clayium.common.unification.OrePrefix
-import com.github.trcdevelopers.clayium.common.unification.material.Material
+import com.github.trcdevelopers.clayium.common.unification.material.EnumMaterial
 import net.minecraft.init.Blocks
 import net.minecraft.item.ItemStack
 
@@ -12,7 +12,7 @@ object CondenserRecipeLoader {
         val registry = CRecipes.CONDENSER
 
         registry.register {
-            input(OrePrefix.DUST, Material.CLAY)
+            input(OrePrefix.DUST, EnumMaterial.CLAY)
             output(ItemStack(Blocks.CLAY))
             cePerTick(ClayEnergy.micro(10))
             duration(3)
