+++ src/main/kotlin/com/github/trcdevelopers/clayium/common/unification/OreDictUnifier.kt
@@ -1,6 +1,6 @@
 package com.github.trcdevelopers.clayium.common.unification
 
-import com.github.trcdevelopers.clayium.common.unification.material.Material
+import com.github.trcdevelopers.clayium.common.unification.material.EnumMaterial
 import com.github.trcdevelopers.clayium.common.unification.stack.UnificationEntry
 import net.minecraft.item.ItemStack
 import net.minecraftforge.oredict.OreDictionary
@@ -14,7 +14,7 @@ object OreDictUnifier {
         }
     }
 
-    fun get(orePrefix: OrePrefix, material: Material, stackSize: Int = 1): ItemStack {
+    fun get(orePrefix: OrePrefix, material: EnumMaterial, stackSize: Int = 1): ItemStack {
         return get(UnificationEntry(orePrefix, material).toString(), stackSize)
     }
 }
