+++ src/main/kotlin/com/github/trcdevelopers/clayium/common/unification/OrePrefix.kt
@@ -1,7 +1,7 @@
 package com.github.trcdevelopers.clayium.common.unification
 
 import com.github.trcdevelopers.clayium.api.util.CUtils
-import com.github.trcdevelopers.clayium.common.unification.material.Material
+import com.github.trcdevelopers.clayium.common.unification.material.EnumMaterial
 import com.github.trcdevelopers.clayium.common.unification.material.MaterialProperty
 import com.google.common.base.CaseFormat
 
@@ -11,26 +11,26 @@ enum class OrePrefix(
 ) {
 
     INGOT("ingot") {
-        override fun doGenerateItem(material: Material) = material.hasProperty<MaterialProperty.Ingot>()
+        override fun doGenerateItem(material: EnumMaterial) = material.hasProperty<MaterialProperty.Ingot>()
     },
     MATTER("matter") {
-        override fun doGenerateItem(material: Material) = material.hasProperty<MaterialProperty.Matter>()
+        override fun doGenerateItem(material: EnumMaterial) = material.hasProperty<MaterialProperty.Matter>()
     },
     DUST("dust") {
-        override fun doGenerateItem(material: Material) = material.hasProperty<MaterialProperty.Dust>()
+        override fun doGenerateItem(material: EnumMaterial) = material.hasProperty<MaterialProperty.Dust>()
     },
     IMPURE_DUST("impureDust") {
-        override fun doGenerateItem(material: Material) = material.hasProperty<MaterialProperty.ImpureDust>()
+        override fun doGenerateItem(material: EnumMaterial) = material.hasProperty<MaterialProperty.ImpureDust>()
     },
     PLATE("plate") {
-        override fun doGenerateItem(material: Material) = material.hasProperty<MaterialProperty.Plate>()
+        override fun doGenerateItem(material: EnumMaterial) = material.hasProperty<MaterialProperty.Plate>()
     },
     LARGE_PLATE("largePlate") {
-        override fun doGenerateItem(material: Material) = material.hasProperty<MaterialProperty.Plate>()
+        override fun doGenerateItem(material: EnumMaterial) = material.hasProperty<MaterialProperty.Plate>()
     },
     ;
 
-    abstract fun doGenerateItem(material: Material): Boolean
+    abstract fun doGenerateItem(material: EnumMaterial): Boolean
 
-    fun concat(material: Material) = "$camel${CaseFormat.LOWER_UNDERSCORE.to(CaseFormat.UPPER_CAMEL, material.materialName)}"
+    fun concat(material: EnumMaterial) = "$camel${CaseFormat.LOWER_UNDERSCORE.to(CaseFormat.UPPER_CAMEL, material.materialName)}"
 }
