+++ src/main/kotlin/com/github/trcdevelopers/clayium/common/unification/material/CMaterials.kt
@@ -1,0 +1,12 @@
+package com.github.trcdevelopers.clayium.common.unification.material
+
+import com.github.trcdevelopers.clayium.api.util.clayiumId
+import com.github.trcdevelopers.clayium.common.clayenergy.ClayEnergy
+
+object CMaterials {
+    val impureSilicon = Material.create(0, clayiumId("impure_silicon")) {
+        colors(0x978F98, 0x533764, 0xA9A5A5)
+        ingot().dust()
+        plate(ClayEnergy.milli(1), 20, tier = 4)
+    }
+}
