+++ src/main/kotlin/com/github/trcdevelopers/clayium/common/unification/material/PropertyKey.kt
@@ -1,11 +1,18 @@
 package com.github.trcdevelopers.clayium.common.unification.material
 
-class PropertyKey<T : MaterialProperty>  {
+data class PropertyKey<T : MaterialProperty>(
+    val name: String,
+)  {
+    @Suppress("UNCHECKED_CAST")
     fun cast(value: MaterialProperty): T {
         return value as T
     }
 
     companion object {
-        val INGOT = PropertyKey<MaterialProperty.Ingot>()
+        val INGOT = PropertyKey<MaterialProperty.Ingot>("ingot")
+        val DUST = PropertyKey<MaterialProperty.Dust>("dust")
+        val MATTER = PropertyKey<MaterialProperty.Matter>("matter")
+        val IMPURE_DUST = PropertyKey<MaterialProperty.ImpureDust>("impureDust")
+        val PLATE = PropertyKey<MaterialProperty.Plate>("plate")
     }
 }
