+++ src/main/kotlin/com/github/trcdevelopers/clayium/common/unification/stack/UnificationEntry.kt
@@ -1,11 +1,11 @@
 package com.github.trcdevelopers.clayium.common.unification.stack
 
 import com.github.trcdevelopers.clayium.common.unification.OrePrefix
-import com.github.trcdevelopers.clayium.common.unification.material.Material
+import com.github.trcdevelopers.clayium.common.unification.material.EnumMaterial
 
 data class UnificationEntry(
     val orePrefix: OrePrefix,
-    val material: Material,
+    val material: EnumMaterial,
 ) {
     override fun toString(): String {
         return "${orePrefix.camel}${material.toUpperCamel()}"
